// src/devices/page/base/types.ts
export type DeviceSelectionProps = {
    capability: TIncyclistCapability
    devices: Array<DeviceSelectionItemProps>
    disabled: boolean,
    isScanning: boolean
    changeForAll: boolean
    canSelectAll: boolean
    onClose: ()=>void
}

export type TConnectState = 'failed' | 'connected' | 'connecting'
export type DeviceSelectionItemProps = {
    connectState?: TConnectState
    isSelected: boolean
    deviceName: string
    value: number,
    interface: string,
    onClick?: (addAll?:boolean)=>void
    onDelete?: ()=>void
}
// ...
export type TIncyclistCapability = 'power' | 'speed' | 'cadence' | 'heartrate' | 'control' | 'app_control'
export type TDisplayCapability ='resistance'|'power'|'heartrate'|'cadence'|'speed'|'controller'

export type PairingExitVia = 'ok' | 'skip' | 'simulate' | 'cancel' | 'app_exit'